+++ KFXAdditions/Classes/UIViewController+KFXAdditions.m
@@ -19,6 +19,33 @@
     [self.presentingViewController dismissViewControllerAnimated:YES completion:nil];
 }
 
+//--------------------------------------------------------
+#pragma mark - Presentation
+//--------------------------------------------------------
+-(BOOL)hasBeenPresentedModally{
+	
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
+    BOOL isModal = ((self.parentViewController && self.parentViewController.modalViewController == self) ||
+                    //or if I have a navigation controller, check if its parent modal view controller is self navigation controller
+                    ( self.navigationController && self.navigationController.parentViewController && self.navigationController.parentViewController.modalViewController == self.navigationController) ||
+                    //or if the parent of my UITabBarController is also a UITabBarController class, then there is no way to do that, except by using a modal presentation
+                    [[[self tabBarController] parentViewController] isKindOfClass:[UITabBarController class]]);
+    
+    //iOS 5+
+    if (!isModal && [self respondsToSelector:@selector(presentingViewController)]) {
+        
+        isModal = ((self.presentingViewController && self.presentingViewController.modalViewController == self) ||
+                   //or if I have a navigation controller, check if its parent modal view controller is self navigation controller
+                   (self.navigationController && self.navigationController.presentingViewController && self.navigationController.presentingViewController.modalViewController == self.navigationController) ||
+                   //or if the parent of my UITabBarController is also a UITabBarController class, then there is no way to do that, except by using a modal presentation
+                   [[[self tabBarController] presentingViewController] isKindOfClass:[UITabBarController class]]);
+        
+    }
+#pragma clang diagnostic pop
+
+    return isModal;
+}
 
 //--------------------------------------------------------
 #pragma mark Alerts